修改前端，增加公益服显示，准备发布

+++ src/main/java/cc/mcyou/bbsindex/data/SeriesData.java
@@ -13,7 +13,7 @@ public class SeriesData {
     String type = "line";
     boolean smooth = true;
     String symbol = "none";
-    List<Integer> data;
+    List<Long> data;
 }
